fix: security configs

+++ src/main/java/com/easymenu/api/security/SecurityProperties.java
@@ -12,5 +12,5 @@ public class SecurityProperties {
     @Value("${security.secret-key}")
     private String secretKey;
     @Value("${security.jwt-ttl}")
-    private Long jwtTtl;
+    private Integer jwtTtl;
 }
